pager: perform incremental reconciliation
Previously the pager cleared out all results to page-in new
reuslts. This was wasteful if the new results almost entirely
overlapped with the old results.

Now, pager will build a sequence of updates to transform the old
results to the new results. It will then perform these in batches.

Even if the old and new results are very long lists, the number of
updates required may be very small if the lists are almost identical.