use localStorage to persist componentState
transit is used to persist Immutable objects.

I also use withRouter for the provider now. I was having a problem
where `this.context.history` would not be up-to-date always in
Provider. So making this a prop provided in the normal way by
react-router presumably ensures that it is always up-to-date.

Note I still rely on the hack where components are *not* connected to
the router; that's why they use out-of-date keys on navigation which
allows the Provider to ignore their clearComponentState request. This
is still a hack!